requirements.txt
discord==2.3.2
requests==2.31.0
flask==3.0.0
python-dotenv==1.0.0